put generate widget code on the page
